Lizard Markings Port (#448)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Port of PR # 673 from Floofstation, adding Head and Chest markings as
well as a large snake tail markng.

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added some head and chest markings for lizards, as well as a
large snake tail marking that animates when wagging.

---------

Co-authored-by: Fansana <116083121+Fansana@users.noreply.github.com>
Co-authored-by: PuroSlavKing <103608145+puroslavking@users.noreply.github.com>