Okay, tested with Dan and finally got it working. I dunno what the crashing issue was with the pens was. But the spaceman figure crash was because of improper tag was used. That was fixed.
Unfortunately didn't fully remedy being able to put in the bag. Dan pointed out it did not have in the storage field a defined maxItemSize field so it was autogenerating a size. So now with that defined, you can stick figures in! Huzzah! Now just wanna work on wallets next, and this should be PR-able.