More small stuff I can do. I had planned on making a cargo crate to replace navigation lights that burn out due to solar flares. Since sometimes, getting the right stupid crystals to make the necessary light color is a chore.
While trying to figure out making a new crate, I found that the Holiday Replacement Light Crate already included red and green lightbulbs in it. But was not available in the cargo console.

Dan taught me that the ``Abstract: True`` parameter is a lot like commenting out a prototype and so that is why it was unavailable.

Rather than enabling it, I had a big brain moment and made a child crate that's renamed for navigation and is then orderable in Cargo. It's functionally identical besides having a better name for it's purpose.